优化超声波传感器和OLED调试输出，调整数据类型并增加新打印功能

+++ US.cpp
@@ -4,11 +4,11 @@
 US::US(){
 }
 
-US::US(int pinTrigger, int pinEcho){
+US::US(uint8_t pinTrigger, uint8_t pinEcho){
   US::init(pinTrigger,pinEcho);
 }
 
-void US::init(int pinTrigger, int pinEcho)
+void US::init(uint8_t pinTrigger, uint8_t pinEcho)
 {
   _pinTrigger = pinTrigger;
   _pinEcho = pinEcho;
